+++ frontend/src/components/question-creation/question-elements/QuestionTextElement.tsx
@@ -3,10 +3,8 @@ import ResizeTextarea from "react-textarea-autosize";
 import { Flex, Text, Textarea } from "@chakra-ui/react";
 
 import QuestionEditorContext from "../../../contexts/QuestionEditorContext";
-import {
-  QuestionElementType,
-  QuestionTextMetadata,
-} from "../../../types/QuestionTypes";
+import { QuestionTextMetadata } from "../../../types/QuestionMetadataTypes";
+import { QuestionElementType } from "../../../types/QuestionTypes";
 import {
   exceedsMaxLength,
   updatedQuestionElement,
