+++ frontend/src/components/question-creation/question-elements/modals/short-answer/AddShortAnswerModal.tsx
@@ -3,10 +3,8 @@ import { useDisclosure } from "@chakra-ui/react";
 import { v4 as uuidv4 } from "uuid";
 
 import QuestionEditorContext from "../../../../../contexts/QuestionEditorContext";
-import {
-  QuestionElementType,
-  ShortAnswerMetadata,
-} from "../../../../../types/QuestionTypes";
+import { ShortAnswerMetadata } from "../../../../../types/QuestionMetadataTypes";
+import { QuestionElementType } from "../../../../../types/QuestionTypes";
 
 import ShortAnswerModal from "./ShortAnswerModal";
