+++ frontend/src/types/QuestionMetadataTypes.ts
@@ -1,0 +1,16 @@
+export interface QuestionTextMetadata {
+  questionText: string;
+}
+
+export interface TextMetadata {
+  text: string;
+}
+
+export interface ShortAnswerMetadata {
+  answer: number;
+}
+
+export interface MultipleChoiceMetadata {
+  options: string[];
+  answerIndex: number;
+}
